+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/listener/internal/QueryChannelsListenerImpl.kt
@@ -55,7 +55,7 @@ internal class QueryChannelsListenerImpl(private val logic: LogicRegistry) : Que
     }
 
     override suspend fun onQueryChannelsResult(result: Result<List<Channel>>, request: QueryChannelsRequest) {
-        // Nothing to do
+        logic.queryChannels(request).onQueryChannelsResult(result, request)
     }
 
     private companion object {