+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   The change was introduced in order to distinguish between initial state when channels are not loaded yet
   and a state when channels don't exist. [#3531](https://github.com/GetStream/stream-chat-android/pull/3531)
 - Adding logs for QuerySort: [3570](https://github.com/GetStream/stream-chat-android/pull/3570)
+- Adding logs for plugin usage, state calls usage and ChannelListView. [3572](https://github.com/GetStream/stream-chat-android/pull/3572)
 
 ### ✅ Added
 - Added `EventHandlingResult.WatchAndAdd` to results returned from `ChatEventHandler`.